Fix bugs with a lot of rulesets.
When a ruleset's id is **greater** than the size of ruleset_table, ruleset_table is extended twice.
If id is equal to size of ruleset_table, it doesn't be extended.
It follows that invalid memory write on ruleset_table->entry[id].

This commit modifies this operation and now ruleset_table is extended when ruleset's id is **greater or equal** to the size of ruleset_table.